dspy_with_pgvector/requirements.txt
dspy-ai
ollama
psycopg2-binary
pgvector-python
InstructorEmbedding
torch
sentence-transformers
beautifulsoup4
tqdm
streamlit